energy_scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.1.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
lxml>=5.0.0
feedparser>=6.0.0
langdetect>=1.0.9
playwright>=1.40.0
cloudscraper>=1.2.71
google-generativeai>=0.8.0
groq>=0.18.0
streamlit>=1.32.0
python-dotenv>=1.0.1
aiohttp>=3.9.0
playwright-stealth>=1.0.632.0
brotli>=1.0.9
brotlicffi>=1.1.0
gspread>=6.0.0
google-auth>=2.27.0
readability-lxml>=0.8.1
duckduckgo-search>=5.0.0
